Remake authorization, remake API, refactor controllers

+++ src/main/java/com/poptsov/marketplace/service/OrderService.java
@@ -9,14 +9,17 @@ import com.poptsov.marketplace.database.repository.UserRepository;
 import com.poptsov.marketplace.dto.OrderCreateDto;
 import com.poptsov.marketplace.dto.OrderReadDto;
 import com.poptsov.marketplace.dto.OrderEditStatusDto;
+import com.poptsov.marketplace.exceptions.AuthorizationException;
 import com.poptsov.marketplace.exceptions.EntityGetException;
 import com.poptsov.marketplace.mapper.OrderCreateMapper;
 import com.poptsov.marketplace.mapper.OrderReadMapper;
+import com.poptsov.marketplace.util.AuthorityCheckUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -28,14 +31,17 @@ public class OrderService {
     private final ShopRepository shopRepository;
     private final OrderReadMapper orderReadMapper;
     private final OrderCreateMapper orderCreateMapper;
+    private final UserService userService;
 
     public List<OrderReadDto> getOrdersByShopId(Integer id) {
+        Integer ownerId = shopRepository.findById(id).orElseThrow(() -> new EntityGetException("Shop not found")).getUser().getId();
+        AuthorityCheckUtil.checkAuthorities(userService.getCurrentUser().getId(), ownerId);
         Shop shop = shopRepository.findById(id).orElseThrow(() -> new EntityGetException("Shop with id: " + id + " not found"));
         return shop.getOrders().stream().map(orderReadMapper::map).collect(Collectors.toList());
     }
 
-    public List<OrderReadDto> getOrdersByUserId(Integer id) {
-        User user = userRepository.findUserById(id).orElseThrow(() -> new EntityGetException("User with id: " + id + " not found"));
+    public List<OrderReadDto> getOrdersByUserId() {
+        User user = userService.getCurrentUser();
        return user.getOrders().stream().map(orderReadMapper::map).collect(Collectors.toList());
     }
 
@@ -45,10 +51,9 @@ public class OrderService {
     }
 
     @Transactional
-    public OrderReadDto createOrder(Integer userId, Integer shopId, OrderCreateDto orderCreateDto) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new EntityGetException("User  not found with id: " + userId));
+    public OrderReadDto createOrder(Integer shopId, OrderCreateDto orderCreateDto) {
 
+        User user = userService.getCurrentUser();
         Shop shop = shopRepository.findById(shopId)
                 .orElseThrow(() -> new EntityGetException("Shop not found with id: " + shopId));
 
@@ -62,7 +67,11 @@ public class OrderService {
 
     @Transactional
     public boolean deleteOrder(Integer id) {
-
+        Integer ownerId = orderRepository.findById(id).orElseThrow(() -> new EntityGetException("Order not found")).getUser().getId();
+        Integer shopOwnerId = orderRepository.findById(id).orElseThrow(() -> new EntityGetException("Order not found")).getShop().getUser().getId();
+        if (!Objects.equals(ownerId, userService.getCurrentUser ().getId()) && !Objects.equals(shopOwnerId, userService.getCurrentUser ().getId())) {
+            throw new AuthorizationException("Вы не можете удалить этот заказ.");
+        }
         if (orderRepository.existsById(id)) {
             orderRepository.deleteOrderById(id);
             return true;
@@ -72,6 +81,8 @@ public class OrderService {
 
     @Transactional
     public OrderReadDto editOrderStatus(Integer id, OrderEditStatusDto orderEditStatusDto) {
+        Integer shopOwnerId = orderRepository.findById(id).orElseThrow(() -> new EntityGetException("Order not found")).getShop().getUser().getId();
+        AuthorityCheckUtil.checkAuthorities(userService.getCurrentUser().getId(), shopOwnerId);
         Order order = orderRepository.findById(id).orElseThrow(() -> new EntityGetException("Order not found"));
         order.setStatus(orderEditStatusDto.getStatus());
         orderRepository.save(order);
